Estoy trabajando en él principalmente en el archivo main.cpp e corrigiendo el código en las librerías que cree para que compile el código - I'm working on it mainly on the main.cpp file and correcting the code in the libraries I create so that it compiles the code

+++ Codigo/solar_tracker_PlatformIO/lib/LedAdvertencia/LedAdvertencia.cpp
@@ -1,7 +1,10 @@
 #include "LedAdvertencia.h"
 
 // Constructor de la clase que recibe el pin para el led de advertencia
-LedAdvertencia::LedAdvertencia(int ledAdvertenciaPIN) : _ledAdvertenciaPIN(ledAdvertenciaPIN) {}
+LedAdvertencia::LedAdvertencia(uint8_t ledAdvertenciaPIN) 
+                                : 
+                                _ledAdvertenciaPIN(ledAdvertenciaPIN) 
+{}
 
 // Inicialización del led de advertencia
 void LedAdvertencia::begin() {
